* win/tclWinSerial.c (SerialCheckProc): corrected time 	calculations to be unsigned. (schroedter)

+++ win/tclWinSerial.c
@@ -79,7 +79,7 @@ typedef struct SerialInfo {
     int readable;               /* flag that the channel is readable */
     int writable;               /* flag that the channel is writable */
     int blockTime;              /* max. blocktime in msec */
-    int lastEventTime;		/* Time in milliseconds since last readable event */
+    unsigned int lastEventTime;	/* Time in milliseconds since last readable event */
 				/* Next readable event only after blockTime */
     DWORD error;                /* pending error code returned by
                                  * ClearCommError() */
@@ -350,7 +350,7 @@ SerialBlockTime(
  *----------------------------------------------------------------------
  */
 
-static int
+static unsigned int
 SerialGetMilliseconds(
     void)
 {
@@ -442,7 +442,7 @@ SerialCheckProc(
     int needEvent;
     ThreadSpecificData *tsdPtr = TCL_TSD_INIT(&dataKey);
     COMSTAT cStat;
-    int time;
+    unsigned int time;
 
     if (!(flags & TCL_FILE_EVENTS)) {
         return;
@@ -493,7 +493,8 @@ SerialCheckProc(
                             (infoPtr->error & SERIAL_READ_ERRORS) ) {
                         infoPtr->readable = 1;
 			time = SerialGetMilliseconds();
-			if ( (time - infoPtr->lastEventTime) >= infoPtr->blockTime) {
+			if ((unsigned int) (time - infoPtr->lastEventTime)
+				>= (unsigned int) infoPtr->blockTime) {
 			    needEvent = 1;
 			    infoPtr->lastEventTime = time;
 			}
